+++ src/components/Navbar.js
@@ -4,7 +4,6 @@ import Container from "react-bootstrap/Container";
 import Navbar from "react-bootstrap/Navbar";
 import Nav from "react-bootstrap/Nav";
 import logo from "../assets/logo.png";
-import SearchBar from "../pages/Recipes/searchbar";
 
 class navbar extends React.Component {
   render() {
